Added new login.tsx. Updated the routing in App.tsx. Removed footer for unauthenicated users
