+++ tabs/src/components/FooterComponent.tsx
@@ -13,7 +13,7 @@ const FooterComponent: React.FC<FooterComponentProps> = ({ hidden }) => {
 }
   return (
     <footer className={styles.footer}>
-      <Link to="/">Home</Link>&nbsp;|&nbsp;
+      <Link to="/leaderboard">Home</Link>&nbsp;|&nbsp;
       <Link to="/users">Users</Link>&nbsp;|&nbsp; 
       <Link to="/rewards">Rewards</Link>&nbsp;|&nbsp;
       <Link to="/wallet">Wallet</Link>&nbsp;|&nbsp;
